flash: check that the target is halted after running operation
If a flash algo operation does not time out, verify that the core
is halted. If not, an exception is raised. Timeout will pick up the
core continuing to run, but the core could have entered lockup or
somehow be sleeping (or even held in reset, depending on the SoC).