flash: mass erase: clean up confusing log messages when using chip erase as fallback
When chip erase is used as the default method of performing a mass
erase, for targets that don't implement Target.mass_erase(), the user
would see both mass and chip erase messages. The mass erase would
always be reported as failing. Both problems are solved, and a single
pair of "Mass erasing device..., Mass erase completed" messages is
logged.