flash: stack canary
If a flash algo dict has an 'end_stack' key with the start address of
the stack region (the end since the stack grows downward), a canary
value will be placed there and checked after each call to the algo.
An overwritten canary causes an error to be logged. For now, no other
action is taken.