flash: increase FLM stack size, stack canary and halt checks, generator script improvements (#1393)
* packs: flash_algo: restructure memory layout; allow larger stack

- Move the flash algo stack to after buffers.
- Between 512 and 8192 bytes will be allocated for the stack depending
on how much memory is available in the given ram region.
- Data buffers start addresses are aligned to 16 bytes.
- Add debug logs for flash algo memory layout.

* flash: stack canary

If a flash algo dict has an 'end_stack' key with the start address of
the stack region (the end since the stack grows downward), a canary
value will be placed there and checked after each call to the algo.
An overwritten canary causes an error to be logged. For now, no other
action is taken.

* flash: check that the target is halted after running operation

If a flash algo operation does not time out, verify that the core
is halted. If not, an exception is raised. Timeout will pick up the
core continuing to run, but the core could have entered lockup or
somehow be sleeping (or even held in reset, depending on the SoC).

* scripts: generate_flash_algo.py: fixes and improvements.

- Increase default stack size to 4 KiB.
- Precompute page buffers.
- Fix type error for "hex" formatted data.
- Don't print copyright warning for --info-only.
- Include zi in SP computation.
- Move stack to after page buffers to match pyocd’s runtime FLM
memory layout.
- Offset ‘{ro,rw,zi}_start’ output dict members with the header size.
- Add ‘end_stack’ to output dict.
- Print all computed values before generating output.
- Print warning if the flash has multiple sector sizes.