packs: device: replace complex flash region creation with simpler integration using CoreSightTarget.create_flash()
Flash regions are built at the same time as the other memory regions,
but the .flm attribute is simply set to the PackFlashAlgo object rather
than fully setting up the algo. Later, CoreSightTarget.create_flash()
and finishes the job. This results in a single implementation for
both DFPs and user-assigned FLM files that supports all required
features.