flash: Add option to prevent resetting of device after programming
This adds a --no-reset command line argument that can be used to
prevent a target being reset after a program operation has finished.
The use-case for such a scenario includes when programming multiple
images whereby one of the images enables security such as disabling
the debugger, without this patch, only the first image would be
flashed.