RTOS: Catch TransferErrors during kernel state checks
In some cases, RTOS data structures needs to be loaded to external
memory space after the program is started or by some kind of
booloader, and these memory spaces might not be available at post
reset event. This patch try to address this issue for some RTOS
ports (RTX5 already has exception rescue routine so is not affected).

Signed-off-by: Yilin Sun <imi415@imi.moe>