pack: flash: place algo at end of RAM with stack going all the way to RAM start
This ensures all RAM can be used for the stack if needed. The downside
is that there isn't much sense to having a stack canary, because if
the stack overflows then a fault will occur (not a bad thing in itself),
assuming the algo RAM start doesn't have accessible memory located
before it.