dependencies: bump pylink-square (#1503)
* [Dependencies] Bump Pylink

Pylink 1.0 has the following breaking change:

@denravonska: Changed .flash() to no longer erase chip on flash; users will now
need to ensure they call .erase() prior to flashing a non-erased region of
flash.

However, pyOCD doesn't use JLinks's flash routines so there is no call
in the codebase to `.flash()` from `pylink`. This makes it safe to allow
`pylink` packages with a major version of 1.

* dependencies: require pylink-square>=1.0,<2.0

---------

Co-authored-by: Chris Reed <flit@me.com>