coresight: dap: revert API change to add DebugPort param to DPConnector constructor
The DebugPort was passed into DPConnector so that a trace log could
be output when reading the DP IDR, thus breaking API compatibility.
This change was reverted and replaced with DPConnector.read_idr()
performing its own trace log with the IDR result. The formatting of
the log message is different, but that's a minor concern.