debug: sequences:  assignment expressions (#1564)
Support C-style assignment expressions, such that variables can be
assigned to in the middle of other expressions. Just as in C, the
value of an assignment expression is the value assigned to the variable.
While this syntax isn't documented in the Open-CMSIS-Pack specification,
it is used in the Keil.LPC800_DFP pack and therefore is implicitly
supported by debug sequences.

The patch replaces the assign_stmt AST node with assign_expr, with the
lowest precedence (same precedence as in C).