core: session: only create one session to access options when there's no current session
If Session.current_session is accessed without a valid current session,
an empty Session is created and retained via a strong reference. This
mostly happens when session options are read prior to the main session
being created.

Since the ._current_session attribute is a weakref, the options session
instance would previously be collected immediately after use, leading
to repeated creation of options sessions in some cases.