RTT_COMMAND: Allow no down-channels
If no down-channel can be found, the target only supports printing RTT
messages.

Before this commit, this target-configuration was only possible to do when
supplying a log file.
Then commands were ignored and only the up-channel was written to the
file.
If no log-file was supplied, we detected the missing down-channels and
returned an error, never starting the viewer_loop so no up-channel
messages were printed on stdout.

With this commit,
instead of returning an error when no down-channels are configured,
we inform the user and start the viewer_loop
with a down-channel (parameter down_chan) with value 'None'.

Inside the viewer_loop we detect the None-value of down_chan, drop the user
supplied command and continue our viewer_loop.